requirements.txt
# Core Flask dependencies
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0

# Configuration and environment
python-dotenv==1.1.1

# Database (MariaDB/MySQL)
SQLAlchemy==2.0.41
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
mariadb==1.1.12

# Security
cryptography==45.0.5
python3-saml==1.16.0
xmlsec==1.3.15

# Development tools
pylint==3.3.7
pylint-flask==0.6
black==25.1.0
isort==6.0.1
mypy==1.16.1

# Testing
pytest==8.4.1
pytest-flask==1.3.0
pytest-cov==6.2.1
coverage==7.9.2

# Pre-commit hooks and security scanning
pre-commit==4.2.0
detect-secrets==1.5.0
bandit==1.8.0
pre_commit_hooks==5.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
